chore: move exceptions to global package (#2)
