+++ src/main/java/flow/extensionblocker/common/global/exception/BlockerNotFoundException.java
@@ -1,4 +1,4 @@
-package flow.extensionblocker.common.exception;
+package flow.extensionblocker.common.global.exception;
 
 import lombok.Getter;
